+++ src/imio/smartweb/core/browser/icons.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+from Products.CMFPlone.browser.icons import IconsView
+
+
+class SmartwebIconsView(IconsView):
+
+    prefix = "smartweb.icon."