+++ src/imio/smartweb/core/contents/blocks/link/content.py
@@ -4,8 +4,7 @@ from imio.smartweb.locales import SmartwebMessageFactory as _
 from plone.dexterity.content import Container
 from plone.namedfile.field import NamedBlobImage
 from zope.interface import implementer
-from zope.schema import Bool
-from zope.schema import Text
+from zope import schema
 from zope.schema import TextLine
 from zope.schema.interfaces import ITextLine
 from plone.app.contenttypes.interfaces import ILink
@@ -23,11 +22,24 @@ class LinkField(TextLine):
 class IBlockLink(ILink):
     """Marker interface and Dexterity Python Schema for BlockLink"""
 
-    title = TextLine(title=_(u"Title"), required=True)
-    description = Text(title=_(u"Description"), required=False)
-    open_in_new_tab = Bool(title=_(u"Open in a new tab"), required=False, default=False)
+    title = schema.TextLine(title=_(u"Title"), required=True)
+
+    description = schema.Text(title=_(u"Description"), required=False)
+
+    open_in_new_tab = schema.Bool(
+        title=_(u"Open in a new tab"), required=False, default=False
+    )
+
     remoteUrl = LinkField(title=_(u"URL"), required=True)
+
     image = NamedBlobImage(title=_(u"Image"), required=False)
+
+    svg_icon = schema.Choice(
+        title=_(u"Icon"),
+        description=_(u"Only used in table view (takes precedence over image)"),
+        source="imio.smartweb.vocabulary.Icons",
+        required=False,
+    )
 
 
 @implementer(IBlockLink)